+++ src/moai-sim/MOAIMesh.cpp
@@ -554,7 +554,7 @@ void MOAIMesh::DrawIndex ( u32 idx, MOAIMeshSpan* span ) {
 					if ( span->mBase == span->mTop ) continue;
 					assert (( span->mBase < span->mTop ) && ( span->mTop <= this->mTotalElements ));
 				
-					gfxMgr.mGfxState.DrawPrims ( this->mPrimType, span->mBase, ( u32 )( span->mTop - span->mBase ));
+					gfxMgr.mGfxState.DrawPrims ( this->mPrimType, ( u32 )span->mBase, ( u32 )( span->mTop - span->mBase ));
 				}
 
 				gfxMgr.mGfxState.SetIndexBuffer ();
@@ -567,7 +567,7 @@ void MOAIMesh::DrawIndex ( u32 idx, MOAIMeshSpan* span ) {
 				if ( span->mBase == span->mTop ) continue;
 				assert (( span->mBase < span->mTop ) && ( span->mTop <= this->mTotalElements ));
 			
-				gfxMgr.mGfxState.DrawPrims ( this->mPrimType, span->mBase, ( u32 )( span->mTop - span->mBase ));
+				gfxMgr.mGfxState.DrawPrims ( this->mPrimType, ( u32 )span->mBase, ( u32 )( span->mTop - span->mBase ));
 			}
 		}
 		gfxMgr.mGfxState.SetVertexArray ();
